folder_cleaner/requirements.txt
# Python dependencies
psutil
google-generativeai
python-dotenv
